Move store-owned tests out of root lib-test
