+++ src/db/agent_wake_tests.rs
@@ -1,15 +1,11 @@
-use super::agent_wake::{AgentWakeAttemptState, AgentWakeState};
-use super::communication::{
-    CommunicationPrincipal, ConversationAccess, NewAgentEndpoint, NewAgentIdentity,
-    NewConversation, NewConversationMessage,
+use super::{
+    AgentWakeEnvelope, AgentWakeState, CommunicationPrincipal, Database, NewAgentEndpoint,
+    NewAgentIdentity, NewConversation, NewConversationMessage,
 };
-use super::Database;
 use crate::agent_wake::{
     dispatch_next_agent_wake, AgentWakeDispatchReport, ContinuationAdapter,
     ContinuationDispatchOutcome, ContinuationPreflight, ContinuationPreflightError,
 };
-use crate::db::AgentWakeEnvelope;
-use rusqlite::params;
 use std::sync::Mutex;
 
 #[derive(Clone)]
@@ -103,7 +99,7 @@ fn attach_wake_endpoint(
     fixture: &Fixture,
     host: &str,
     key: &str,
-) -> super::communication::AgentEndpointRecord {
+) -> super::AgentEndpointRecord {
     db.attach_agent_endpoint(
         &fixture.owner,
         NewAgentEndpoint {
@@ -144,91 +140,6 @@ fn queued_delivery_ids(db: &Database, agent_id: &str) -> Vec<String> {
         .unwrap()
         .collect::<Result<Vec<String>, _>>()
         .unwrap()
-}
-
-#[test]
-fn withdrawing_wake_capability_reconciles_exact_endpoint_attempts() {
-    let temp = tempfile::tempdir().unwrap();
-    let db = Database::open(&temp.path().join("wake-capability-withdrawal.db")).unwrap();
-    let fixture = create_fixture(&db, 'c');
-    post_to_receiver(&db, &fixture, "withdraw capability", "withdraw-message");
-    let endpoint = attach_wake_endpoint(&db, &fixture, "host", "withdraw-endpoint");
-    let wake_id = wake_id_for(&db, &fixture.receiver_agent_id);
-
-    let _claimed = db
-        .claim_next_agent_wake(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &endpoint.endpoint_id,
-            endpoint.controller_generation,
-            "host_adapter",
-        )
-        .unwrap()
-        .unwrap();
-    let withdrawn = db
-        .set_agent_endpoint_wake_capability(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &endpoint.endpoint_id,
-            endpoint.controller_generation,
-            false,
-        )
-        .unwrap();
-    assert!(!withdrawn.wake_capable);
-    assert_eq!(
-        db.agent_wake(&wake_id).unwrap().unwrap().state,
-        AgentWakeState::Pending
-    );
-    assert_eq!(
-        db.agent_wake_attempts(&wake_id).unwrap()[0].state,
-        AgentWakeAttemptState::Revoked
-    );
-
-    db.set_agent_endpoint_wake_capability(
-        &fixture.owner,
-        &fixture.receiver_agent_id,
-        &endpoint.endpoint_id,
-        endpoint.controller_generation,
-        true,
-    )
-    .unwrap();
-    let claimed = db
-        .claim_next_agent_wake(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &endpoint.endpoint_id,
-            endpoint.controller_generation,
-            "host_adapter",
-        )
-        .unwrap()
-        .unwrap();
-    db.prepare_agent_wake_dispatch(
-        &fixture.owner,
-        &fixture.receiver_agent_id,
-        &endpoint.endpoint_id,
-        endpoint.controller_generation,
-        &wake_id,
-        &claimed.attempt.attempt_id,
-        &claimed.claim_fence,
-        &claimed.consume_token,
-    )
-    .unwrap();
-    db.set_agent_endpoint_wake_capability(
-        &fixture.owner,
-        &fixture.receiver_agent_id,
-        &endpoint.endpoint_id,
-        endpoint.controller_generation,
-        false,
-    )
-    .unwrap();
-    assert_eq!(
-        db.agent_wake(&wake_id).unwrap().unwrap().state,
-        AgentWakeState::DeliveryUnknown
-    );
-    assert_eq!(
-        db.agent_wake_attempts(&wake_id).unwrap()[1].state,
-        AgentWakeAttemptState::DeliveryUnknown
-    );
 }
 
 #[derive(Debug)]
@@ -278,251 +189,6 @@ impl ContinuationAdapter for FakeContinuationAdapter {
         self.envelopes.lock().unwrap().push(envelope.clone());
         self.outcome
     }
-}
-
-#[test]
-fn offline_fifty_message_burst_preserves_facts_and_coalesces_wake() {
-    let temp = tempfile::tempdir().unwrap();
-    let db = Database::open(&temp.path().join("coalescing.db")).unwrap();
-    let fixture = create_fixture(&db, '1');
-
-    for index in 0..50 {
-        post_to_receiver(
-            &db,
-            &fixture,
-            &format!("offline durable message {index}"),
-            &format!("burst-message-{index}"),
-        );
-    }
-
-    let (message_count, delivery_count, wake_count): (i64, i64, i64) = {
-        let conn = db.conn_for_tests();
-        (
-            conn.query_row("SELECT COUNT(*) FROM wc_conversation_messages", [], |row| {
-                row.get(0)
-            })
-            .unwrap(),
-            conn.query_row("SELECT COUNT(*) FROM wc_agent_deliveries", [], |row| {
-                row.get(0)
-            })
-            .unwrap(),
-            conn.query_row("SELECT COUNT(*) FROM wc_agent_wakes", [], |row| row.get(0))
-                .unwrap(),
-        )
-    };
-    assert_eq!((message_count, delivery_count, wake_count), (50, 50, 1));
-
-    let wake_id = wake_id_for(&db, &fixture.receiver_agent_id);
-    let wake = db.agent_wake(&wake_id).unwrap().unwrap();
-    assert_eq!(wake.state, AgentWakeState::Pending);
-    assert_eq!(wake.queued_delivery_count_snapshot, 50);
-    assert!(wake.inbox_high_watermark >= 50);
-    assert_ne!(
-        wake.first_triggering_delivery_id,
-        wake.latest_triggering_delivery_id
-    );
-
-    let agent = db
-        .list_agent_identities(&fixture.owner, Some(&fixture.receiver_agent_id), 0, 10)
-        .unwrap()
-        .agents
-        .pop()
-        .unwrap();
-    assert_eq!(agent.active_endpoint_count, 0);
-    assert_eq!(agent.queued_delivery_count, 50);
-    assert_eq!(agent.unresolved_wake_count, 1);
-    assert_eq!(agent.latest_wake_id.as_deref(), Some(wake_id.as_str()));
-    assert_eq!(agent.latest_wake_state.as_deref(), Some("pending"));
-
-    let endpoint = attach_wake_endpoint(&db, &fixture, "fake-host", "burst-endpoint");
-    let inbox = db
-        .list_agent_inbox(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &endpoint.endpoint_id,
-            endpoint.controller_generation,
-            0,
-            100,
-        )
-        .unwrap();
-    assert_eq!(inbox.total_queued_count, 50);
-    assert_eq!(inbox.deliveries.len(), 50);
-
-    let first_delivery_id = inbox.deliveries[0].delivery_id.clone();
-    db.consume_agent_deliveries(
-        &fixture.owner,
-        &fixture.receiver_agent_id,
-        &endpoint.endpoint_id,
-        endpoint.controller_generation,
-        vec![first_delivery_id],
-    )
-    .unwrap();
-    assert_eq!(
-        db.agent_wake(&wake_id).unwrap().unwrap().state,
-        AgentWakeState::Pending,
-        "Delivery consume must not consume or rewrite the logical Wake"
-    );
-    let transcript = db
-        .read_conversation(
-            &fixture.owner,
-            &ConversationAccess::Human,
-            &fixture.conversation_id,
-            0,
-            100,
-        )
-        .unwrap();
-    assert_eq!(transcript.messages.len(), 50);
-    assert_eq!(transcript.conversation.message_count, 50);
-}
-
-#[test]
-fn replacement_generation_fences_old_claim_dispatch_wake_consume_and_inbox_consume() {
-    let temp = tempfile::tempdir().unwrap();
-    let db = Database::open(&temp.path().join("generation.db")).unwrap();
-    let fixture = create_fixture(&db, '2');
-    post_to_receiver(
-        &db,
-        &fixture,
-        "generation-fenced work",
-        "generation-message",
-    );
-    let delivery_id = queued_delivery_ids(&db, &fixture.receiver_agent_id)
-        .pop()
-        .unwrap();
-
-    let generation_one = attach_wake_endpoint(&db, &fixture, "host-one", "endpoint-one");
-    assert_eq!(generation_one.controller_generation, 1);
-    let claim_one = db
-        .claim_next_agent_wake(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &generation_one.endpoint_id,
-            generation_one.controller_generation,
-            "deterministic_fake",
-        )
-        .unwrap()
-        .unwrap();
-
-    let generation_two = attach_wake_endpoint(&db, &fixture, "host-two", "endpoint-two");
-    assert_eq!(generation_two.controller_generation, 2);
-    let lifecycle: String = db
-        .conn_for_tests()
-        .query_row(
-            "SELECT lifecycle FROM wc_agent_endpoints WHERE endpoint_id = ?1",
-            [&generation_one.endpoint_id],
-            |row| row.get(0),
-        )
-        .unwrap();
-    assert_eq!(lifecycle, "expired");
-    assert_eq!(
-        db.agent_wake(&claim_one.wake.wake_id)
-            .unwrap()
-            .unwrap()
-            .state,
-        AgentWakeState::Pending
-    );
-    assert_eq!(
-        db.agent_wake_attempts(&claim_one.wake.wake_id).unwrap()[0].state,
-        AgentWakeAttemptState::Revoked
-    );
-
-    let stale_claim = db
-        .claim_next_agent_wake(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &generation_one.endpoint_id,
-            generation_one.controller_generation,
-            "deterministic_fake",
-        )
-        .unwrap_err();
-    assert_eq!(stale_claim.code(), "endpoint_expired");
-    assert_eq!(
-        db.prepare_agent_wake_dispatch(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &generation_one.endpoint_id,
-            generation_one.controller_generation,
-            &claim_one.wake.wake_id,
-            &claim_one.attempt.attempt_id,
-            &claim_one.claim_fence,
-            &claim_one.consume_token,
-        )
-        .unwrap_err()
-        .code(),
-        "endpoint_expired"
-    );
-    assert_eq!(
-        db.consume_agent_wake(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &generation_one.endpoint_id,
-            generation_one.controller_generation,
-            &claim_one.wake.wake_id,
-            &claim_one.consume_token,
-        )
-        .unwrap_err()
-        .code(),
-        "endpoint_expired"
-    );
-    assert_eq!(
-        db.consume_agent_deliveries(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &generation_one.endpoint_id,
-            generation_one.controller_generation,
-            vec![delivery_id],
-        )
-        .unwrap_err()
-        .code(),
-        "endpoint_expired"
-    );
-
-    let claim_two = db
-        .claim_next_agent_wake(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &generation_two.endpoint_id,
-            generation_two.controller_generation,
-            "deterministic_fake",
-        )
-        .unwrap()
-        .unwrap();
-    assert_eq!(claim_two.wake.wake_id, claim_one.wake.wake_id);
-    db.prepare_agent_wake_dispatch(
-        &fixture.owner,
-        &fixture.receiver_agent_id,
-        &generation_two.endpoint_id,
-        generation_two.controller_generation,
-        &claim_two.wake.wake_id,
-        &claim_two.attempt.attempt_id,
-        &claim_two.claim_fence,
-        &claim_two.consume_token,
-    )
-    .unwrap();
-    let generation_three = attach_wake_endpoint(&db, &fixture, "host-three", "endpoint-three");
-    assert_eq!(generation_three.controller_generation, 3);
-    assert_eq!(
-        db.verify_agent_wake_dispatch_binding(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &generation_two.endpoint_id,
-            generation_two.controller_generation,
-            &claim_two.wake.wake_id,
-            &claim_two.attempt.attempt_id,
-            &claim_two.claim_fence,
-        )
-        .unwrap_err()
-        .code(),
-        "endpoint_expired",
-        "a callback reaching the controller after replacement is fenced before Host invocation"
-    );
-    assert_eq!(
-        db.agent_wake(&claim_two.wake.wake_id)
-            .unwrap()
-            .unwrap()
-            .state,
-        AgentWakeState::DeliveryUnknown
-    );
 }
 
 #[test]
@@ -789,162 +455,4 @@ fn dispatch_uncertainty_blocks_duplicate_and_preserves_successor_across_restart(
         reopened.agent_wake_attempts(&first_wake_id).unwrap().len(),
         1
     );
-}
-
-#[test]
-fn generic_reopen_preserves_live_pre_dispatch_claim_without_takeover() {
-    let temp = tempfile::tempdir().unwrap();
-    let path = temp.path().join("offline-restart.db");
-    let db = Database::open(&path).unwrap();
-    let fixture = create_fixture(&db, '5');
-    post_to_receiver(
-        &db,
-        &fixture,
-        "offline before attachment",
-        "offline-message",
-    );
-    let wake_id = wake_id_for(&db, &fixture.receiver_agent_id);
-    let endpoint = attach_wake_endpoint(&db, &fixture, "restart-host", "restart-endpoint");
-    let first_claim = db
-        .claim_next_agent_wake(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &endpoint.endpoint_id,
-            endpoint.controller_generation,
-            "deterministic_fake",
-        )
-        .unwrap()
-        .unwrap();
-    assert_eq!(first_claim.wake.wake_id, wake_id);
-
-    let reopened = Database::open(&path).unwrap();
-    assert_eq!(
-        queued_delivery_ids(&reopened, &fixture.receiver_agent_id).len(),
-        1
-    );
-    assert_eq!(
-        reopened.agent_wake(&wake_id).unwrap().unwrap().state,
-        AgentWakeState::Claimed,
-        "opening the database must not assert that a live Wake owner died"
-    );
-    let attempts = reopened.agent_wake_attempts(&wake_id).unwrap();
-    assert_eq!(attempts.len(), 1);
-    assert_eq!(attempts[0].state, AgentWakeAttemptState::Claimed);
-    assert_eq!(attempts[0].attempt_id, first_claim.attempt.attempt_id);
-}
-
-#[test]
-fn generic_reopen_preserves_prepared_dispatch_without_inventing_takeover() {
-    let temp = tempfile::tempdir().unwrap();
-    let path = temp.path().join("prepared-restart.db");
-    let db = Database::open(&path).unwrap();
-    let fixture = create_fixture(&db, '6');
-    post_to_receiver(&db, &fixture, "prepared before restart", "prepared-message");
-    let endpoint = attach_wake_endpoint(&db, &fixture, "prepared-host", "prepared-endpoint");
-    let claim = db
-        .claim_next_agent_wake(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &endpoint.endpoint_id,
-            endpoint.controller_generation,
-            "deterministic_fake",
-        )
-        .unwrap()
-        .unwrap();
-    db.prepare_agent_wake_dispatch(
-        &fixture.owner,
-        &fixture.receiver_agent_id,
-        &endpoint.endpoint_id,
-        endpoint.controller_generation,
-        &claim.wake.wake_id,
-        &claim.attempt.attempt_id,
-        &claim.claim_fence,
-        &claim.consume_token,
-    )
-    .unwrap();
-
-    let reopened = Database::open(&path).unwrap();
-    assert_eq!(
-        reopened
-            .agent_wake(&claim.wake.wake_id)
-            .unwrap()
-            .unwrap()
-            .state,
-        AgentWakeState::Prepared
-    );
-    assert_eq!(
-        reopened.agent_wake_attempts(&claim.wake.wake_id).unwrap()[0].state,
-        AgentWakeAttemptState::Prepared
-    );
-    assert!(
-        reopened
-            .claim_next_agent_wake(
-                &fixture.owner,
-                &fixture.receiver_agent_id,
-                &endpoint.endpoint_id,
-                endpoint.controller_generation,
-                "deterministic_fake",
-            )
-            .unwrap()
-            .is_none(),
-        "a prepared Wake remains fenced until explicit authoritative recovery"
-    );
-}
-
-#[test]
-fn wake_storage_keeps_stable_refs_and_hashes_without_communication_payload() {
-    let temp = tempfile::tempdir().unwrap();
-    let db = Database::open(&temp.path().join("privacy.db")).unwrap();
-    let fixture = create_fixture(&db, '7');
-    let body = "private-body-not-a-wake-fact";
-    post_to_receiver(&db, &fixture, body, "private-idempotency-key");
-    let endpoint = attach_wake_endpoint(&db, &fixture, "privacy-host", "privacy-endpoint");
-    let claim = db
-        .claim_next_agent_wake(
-            &fixture.owner,
-            &fixture.receiver_agent_id,
-            &endpoint.endpoint_id,
-            endpoint.controller_generation,
-            "deterministic_fake",
-        )
-        .unwrap()
-        .unwrap();
-
-    let columns: Vec<String> = {
-        let conn = db.conn_for_tests();
-        let mut statement = conn.prepare("PRAGMA table_info(wc_agent_wakes)").unwrap();
-        statement
-            .query_map([], |row| row.get(1))
-            .unwrap()
-            .collect::<Result<Vec<_>, _>>()
-            .unwrap()
-    };
-    for forbidden in [
-        "body",
-        "description",
-        "specialty_labels",
-        "principal_digest",
-        "idempotency_key",
-        "wake_payload",
-        "credential",
-    ] {
-        assert!(!columns.iter().any(|column| column == forbidden));
-    }
-
-    let (claim_fence_hash, consume_token_hash): (String, String) = db
-        .conn_for_tests()
-        .query_row(
-            "SELECT claim_fence_hash, consume_token_hash
-             FROM wc_agent_wake_attempts WHERE attempt_id = ?1",
-            params![claim.attempt.attempt_id],
-            |row| Ok((row.get(0)?, row.get(1)?)),
-        )
-        .unwrap();
-    assert_ne!(claim_fence_hash, claim.claim_fence);
-    assert_ne!(consume_token_hash, claim.consume_token);
-    assert!(!claim_fence_hash.contains(&claim.claim_fence));
-    assert!(!consume_token_hash.contains(&claim.consume_token));
-    assert_eq!(claim.wake.trigger_kind, "inbox_changed");
-    assert_eq!(claim.wake.state.as_str(), "claimed");
-    assert!(!format!("{:?}", claim.wake).contains(body));
 }
